% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2.R
\name{plot2}
\alias{plot2}
\title{A bar chart
The function creates a bar chart rapresenting the mother'sweight in pounds at last menstrual period}
\usage{
plot2()
}
\value{

}
\description{
A bar chart
The function creates a bar chart rapresenting the mother'sweight in pounds at last menstrual period
}
