% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_data}
\alias{read_data}
\title{Acquisition of dataset}
\usage{
read_data(MASS, birthwt)
}
\arguments{
\item{MASS}{This is the package where dataframe come from}

\item{birthwt}{selected dataframe}
}
\value{
data
}
\description{
Acquisition of dataset
}
\examples{
read_data()
}
