% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regfunction.R
\name{regfunction}
\alias{regfunction}
\title{Linear Regression}
\usage{
regfunction()
}
\arguments{
\item{birthwt}{set of 189 rown and 10 columns}
}
\description{
This function runs two different multiple linear regression: the firts one is a "total" multiple linear
regression with all the parameters; the second one is a multiple linear regression with the parameters that are tested significat from the fist regression
}
